requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
blessings==1.7
bootstrap4==0.1.0
bpython==0.22.1
Brlapi==0.8.2
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
cryptography==3.3.2
cupshelpers==1.0
curtsies==0.3.10
cwcwidth==0.1.6
dbus-python==1.2.16
decorator==5.1.1
distro==1.5.0
distro-info==1.0
Django==4.0.4
django-bootstrap4==22.2
django-cleanup==6.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
easy-thumbnails==2.8.3
executing==0.8.3
greenlet==1.1.2
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
ipython==8.3.0
jedi==0.18.1
loguru==0.6.0
louis==3.16.0
lxml==4.6.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mutagen==1.45.1
numpy==1.19.5
olefile==0.46
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.29
psutil==5.8.0
pure-eval==0.2.2
pycairo==1.16.2
pycryptodomex==3.15.0
pycups==2.0.1
pycurl==7.43.0.6
pygame==2.1.2
Pygments==2.12.0
PyGObject==3.38.0
pyinotify==0.9.6
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
python-pam==1.8.4
python-xapp==2.0.2
python-xlib==0.29
pytz==2021.1
pyxdg==0.27
reportbug==7.10.3+deb11u1
reportlab==3.5.59
requests==2.25.1
scour==0.38.2
setproctitle==1.2.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
tinycss==0.4
tinycss2==1.0.2
traitlets==5.2.1.post0
typing-extensions==4.2.0
unattended-upgrades==0.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
xdg==5
yt-dlp==2022.6.29
